LayoutTests/media/progress-event.html

<html>
    <head>

        <script src=media-file.js></script>
        <script src=video-test.js></script>
        <script>
            function testOnLoadStart()
            {
                consoleWrite("EVENT(" + event.type + ")");
                testExpected("!event.lengthComputable", true);
                testExpected("event.total", 0);
                testExpected("event.loaded", 0);
                endTest();
            }
        
            function start()
            {
                setSrcByTagName("video", findMediaFile("video", "content/test"));
                findMediaElement();
            }
        </script>
    </head>

    <body onload="start()">

        <video controls onloadstart="testOnLoadStart()" >
        </video>

    </body>
</html>
