LayoutTests/media/progress-event-expected.txt

EVENT(loadstart)
EXPECTED (!event.lengthComputable == 'true') OK
EXPECTED (event.total == '0') OK
EXPECTED (event.loaded == '0') OK
END OF TEST
